Programming_Assignment_1

Module parse

Module to parse input files for PA1.

Parses configurations of match, mismatch, gap, and consecutive gap penalties. Parses genome strands contained in the .fasta format.

Structs

Params

A configuration struct to specify params used in the Global and Local Alignment Algorithms.

Functions

extract_sequences

Function to extract DNA sequences in the .fasta file format. Return: Collection of each sequence, as a String.

flatten_into_sequences 🔒

Takes lines from a file and flattens them into DNA sequences, ignoring non-sequence lines. Also parses the names of the sequences.

get_config

Extracts information from the file with the inputted name into a Param struct instance.

split_lines 🔒

Takes an input file name and returns an iterator over each line in the file.